requirements.txt
Pillow==9.2.0
numpy==1.23.2
requests==2.28.1
scipy==1.9.1
